Can now 'create' a clock or weather widget
